building out journey page
