+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/reportWebVitals.js":"1","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/landingpage/LandingPage.js":"2"},{"size":362,"mtime":1613948017515,"results":"3","hashOfConfig":"4"},{"size":860,"mtime":1613948017513,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1j0ixyz",{"filePath":"8","messages":"9","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/reportWebVitals.js",[],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/landingpage/LandingPage.js",["10"],{"ruleId":"11","severity":1,"message":"12","line":23,"column":9,"nodeType":"13","endLine":23,"endColumn":50},"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement"]
+[{"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/reportWebVitals.js":"1","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/landingpage/LandingPage.js":"2","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/index.js":"3","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/context/AppContext.js":"4","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/App.js":"5","/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/journey/Journey.js":"6"},{"size":362,"mtime":1613948017515,"results":"7","hashOfConfig":"8"},{"size":860,"mtime":1613948017513,"results":"9","hashOfConfig":"8"},{"size":548,"mtime":1613948017512,"results":"10","hashOfConfig":"8"},{"size":838,"mtime":1613948017512,"results":"11","hashOfConfig":"8"},{"size":672,"mtime":1614010589093,"results":"12","hashOfConfig":"8"},{"size":783,"mtime":1614011192627,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1j0ixyz",{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/reportWebVitals.js",[],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/landingpage/LandingPage.js",["26"],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/index.js",[],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/context/AppContext.js",[],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/App.js",[],"/Users/tracilevine/Desktop/MVP-projects/Marketplace/client/src/pages/journey/Journey.js",[],{"ruleId":"27","severity":1,"message":"28","line":23,"column":9,"nodeType":"29","endLine":23,"endColumn":50},"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement"]
