+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Login from './pages/login/Login.js';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import LandingPage from './pages/landingpage/LandingPage';
-import Who from './pages/who/Who';
+import Journey from './pages/journey/Journey';
 import { AppContextProvider } from './context/AppContext';
 
 const App = () => {
@@ -12,7 +12,7 @@ const App = () => {
         <Switch>
           <Route exact path='/' component={LandingPage} />
           <Route exact path='/login' component={Login} />
-          <Route exact path='/who' component={Who} />
+          <Route exact path='/journey' component={Journey} />
         </Switch>
       </BrowserRouter>
     </AppContextProvider>
